requirements.txt
Flask==2.0.1
plotly==5.1.0
pandas==1.3.0
mysql-connector-python==8.0.25
